bug fix for year/date post management

+++ src/app/admin/api/post/update/route.ts
@@ -1,5 +1,3 @@
-import { parse } from "date-fns";
-
 import {
   deleteFile,
   getPostDir,
@@ -29,7 +27,7 @@ export async function POST(req: Request) {
       const mainFilenameToDelete = formData.get(
         "mainFilenameToDelete",
       ) as string;
-      if (mainFilenameToDelete !== "") {
+      if (mainFilenameToDelete) {
         if (deleteFile(dir, mainFilenameToDelete)) {
           await prisma.postImage.delete({
             where: { filename: mainFilenameToDelete },
@@ -37,11 +35,13 @@ export async function POST(req: Request) {
         }
       }
       const filenamesToDelete = formData.get("filenamesToDelete") as string;
-      for await (const filename of filenamesToDelete.split(",")) {
-        if (deleteFile(dir, filename)) {
-          await prisma.postImage.delete({
-            where: { filename },
-          });
+      if (filenamesToDelete) {
+        for await (const filename of filenamesToDelete.split(",")) {
+          if (deleteFile(dir, filename)) {
+            await prisma.postImage.delete({
+              where: { filename },
+            });
+          }
         }
       }
 
@@ -90,7 +90,7 @@ export async function POST(req: Request) {
         where: { id },
         data: {
           title,
-          date: parse(formData.get("date") as string, "yyyy", new Date()),
+          date: new Date(Number(formData.get("date")), 1),
           text: formData.get("text") as string,
           images: {
             create: images,
